requirements.txt
tensorflow>=2.13.0
tensorflow-model-optimization>=0.7.4
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
jupyter>=1.0.0
tqdm>=4.65.0
psutil>=5.9.0